Sync 2 steppers without acceleration

+++ SmoothMotion.h
@@ -78,7 +78,7 @@ class SmoothMotion {
     uint32_t m_id;
     uint32_t m_minWaitPulse;
     uint8_t m_statePulse;
-    uint32_t m_numWaitPulse;
+    float m_numWaitPulse;
     uint32_t m_pulseCount;
 
     uint8_t m_stateControl;
